+++ app/home/constants.js
@@ -1,3 +1,5 @@
-const CRAWL_LIMIT = 25 // Number of shows to crawl
-const CRAWL_TV_SHOWS_LIMIT = 25 // Number of TVshows to crawl
-const CRAWL_TV_SHOWS_FAVORITES_LIMIT = 5 // Number of links with the latest episodes for each favorite show
+/* eslint-disable no-unused-vars */
+const CRAWL_LIMIT = 40 // Number of shows to crawl
+const CRAWL_TV_SHOWS_LIMIT = 30 // Number of TVshows to crawl
+const CRAWL_TV_SHOWS_FAVORITES_LIMIT = 6 // Number of links with the latest episodes for each favorite show
+const SEARCHER_MAX_RESULTS = 3 // Max results on searching shows